add winner component
