+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
 import './styles/registerCreatePlayer.css';
 import CreatePlayer from './components/CreatePlayer';
-
+import ViewWinner from './components/Winner';
 import BoardClass from './classes/Board';
 import { FormEvent, useEffect, useState } from 'react';
 import PlayerClass from './classes/Player';
@@ -239,61 +239,68 @@ function App() {
   //   );
   // };
 
-  const handleResetGame = () => {
-    board.resetBoard();
-  };
+  // const handleResetGame = () => {
+  //   board.resetBoard();
+  // };
 
-  const handleNewGame = () => {
-    setState('board', new BoardClass(() => setState()));
-    setState('playerRed', null);
-    setState('playerYellow', null);
-    window.location.reload();
-  };
-  const ViewWinner = () => {
-    return (
-      <div className='gameover-info'>
-        {board.winner ? (
-          <>
-            <h2>The winner is</h2>
+  // const handleNewGame = () => {
+  //   setState('board', new BoardClass(() => setState()));
+  //   setState('playerRed', null);
+  //   setState('playerYellow', null);
+  //   window.location.reload();
+  // };
+  // const ViewWinner = () => {
+  //   return (
+  //     <div className='gameover-info'>
+  //       {board.winner ? (
+  //         <>
+  //           <h2>The winner is</h2>
 
-            <div className='winner-display'>
-              {' '}
-              <h3 className={state.board.winner === '' ? '' : state.board.winner === 'Red' ? 'red-text' : 'yellow-text'}>
-                {' '}
-                {state.board.winner}
-              </h3>
-              <h3 className={state.board.winner === '' ? '' : state.board.winner === 'Red' ? 'red-text' : 'yellow-text'}>
-                {state.board.winner === 'Red' ? playerRed!.name : playerYellow!.name}
-              </h3>
-            </div>
-          </>
-        ) : (
-          <>
-            {' '}
-            <h2>It's a tie.</h2>
-          </>
-        )}
+  //           <div className='winner-display'>
+  //             {' '}
+  //             <h3 className={state.board.winner === '' ? '' : state.board.winner === 'Red' ? 'red-text' : 'yellow-text'}>
+  //               {' '}
+  //               {state.board.winner}
+  //             </h3>
+  //             <h3 className={state.board.winner === '' ? '' : state.board.winner === 'Red' ? 'red-text' : 'yellow-text'}>
+  //               {state.board.winner === 'Red' ? playerRed!.name : playerYellow!.name}
+  //             </h3>
+  //           </div>
+  //         </>
+  //       ) : (
+  //         <>
+  //           {' '}
+  //           <h2>It's a tie.</h2>
+  //         </>
+  //       )}
 
-        <button
-          className='reset-btn btn'
-          onClick={handleResetGame}
-        >
-          Reset Game
-        </button>
-        <button
-          className='reset-btn btn'
-          onClick={handleNewGame}
-        >
-          New Game
-        </button>
-      </div>
-    );
-  };
+  //       <button
+  //         className='reset-btn btn'
+  //         onClick={handleResetGame}
+  //       >
+  //         Reset Game
+  //       </button>
+  //       <button
+  //         className='reset-btn btn'
+  //         onClick={handleNewGame}
+  //       >
+  //         New Game
+  //       </button>
+  //     </div>
+  //   );
+  // };
 
   return (
     <>
       {!playerRed || !playerYellow ? <CreatePlayer state={state} /> : board.render(playerRed, playerYellow)}
-      {!board.gameOver ? <div className='game-currentplayer'></div> : <ViewWinner />}
+      {!board.gameOver ? (
+        <div className='game-currentplayer'></div>
+      ) : (
+        <ViewWinner
+          state={state}
+          setState={setState}
+        />
+      )}
     </>
   );
 }
